project/vuetify-pj-order_control/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "openai",
    "openai-whisper",
    "pykakasi",
    "python-dotenv",
    "python-multipart",
    "rapidfuzz",
    "uvicorn"
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
